# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Game API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/games/{game}/mods": {
            "get": {
                "tags": [
                    "Mods"
                ],
                "summary": "Get mods by game",
                "description": "Get a list of mods for a specific game.",
                "operationId": "getModsByGame",
                "parameters": [
                    {
                        "name": "game",
                        "in": "path",
                        "description": "ID of the game",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Mod"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Game not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Mods"
                ],
                "summary": "Create a new mod for a game",
                "description": "Create a new mod for a specific game with the provided name.",
                "operationId": "createMod",
                "parameters": [
                    {
                        "name": "game",
                        "in": "path",
                        "description": "ID of the game",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Mod name",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name of the mod",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mod created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Mod"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "Game not found"
                    }
                }
            }
        },
        "/api/games/{game}/mods/{mod}": {
            "get": {
                "tags": [
                    "Mods"
                ],
                "summary": "Get mod by ID",
                "description": "Get details of a specific mod by ID for a game.",
                "operationId": "getModById",
                "parameters": [
                    {
                        "name": "game",
                        "in": "path",
                        "description": "ID of the game",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "mod",
                        "in": "path",
                        "description": "ID of the mod",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Mod"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mod or game not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Mods"
                ],
                "summary": "Update a mod for a game",
                "description": "Update an existing mod for a specific game by ID.",
                "operationId": "updateMod",
                "parameters": [
                    {
                        "name": "game",
                        "in": "path",
                        "description": "ID of the game",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "mod",
                        "in": "path",
                        "description": "ID of the mod",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated mod data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "New name of the mod",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mod updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Mod"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "Mod or game not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Mods"
                ],
                "summary": "Delete a mod for a game",
                "description": "Delete an existing mod for a specific game by ID.",
                "operationId": "deleteMod",
                "parameters": [
                    {
                        "name": "game",
                        "in": "path",
                        "description": "ID of the game",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "mod",
                        "in": "path",
                        "description": "ID of the mod",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Mod deleted successfully"
                    },
                    "404": {
                        "description": "Mod or game not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Mod": {
                "title": "Mod",
                "description": "Mod object",
                "properties": {
                    "id": {
                        "description": "ID of the mod",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Name of the mod",
                        "type": "string"
                    },
                    "user_id": {
                        "description": "ID of the user who created the mod",
                        "type": "integer"
                    },
                    "game_id": {
                        "description": "ID of the game who created the mod",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "Date and time when the mod was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date and time when the mod was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    }
}